+++ src/AudioSensor.h
@@ -11,7 +11,7 @@ class AudioSensor {
         static const uint8_t DEFAULT_MIC_GAIN = 40;
         static const uint8_t DEFAULT_LINE_IN_LEVEL = 5;
         static const uint8_t MAX_MEMORY_BLOCKS = 30;
-        static constexpr float LOWPASS_FREQUENCY = 250;
+        static constexpr double LOWPASS_FREQUENCY = 250;
 
         AudioInputI2S audioInput;
 
@@ -81,8 +81,8 @@ class AudioSensor {
         uint8_t getLineInLevel();
         void increaseLineInLevel();
         void decreaseLineInLevel();
-        float getNormalizedMicGain();
-        float getNormalizedLineInLevel();
+        double getNormalizedMicGain();
+        double getNormalizedLineInLevel();
         void printStats();
 };
