+++ src/EllipticMotion.cpp
@@ -31,32 +31,32 @@ EllipticMotion& EllipticMotion::setSaturation(uint8_t saturation) {
     return *this;
 }
 
-EllipticMotion& EllipticMotion::setCenter(float center) {
+EllipticMotion& EllipticMotion::setCenter(double center) {
     this->center = center;
     return *this;    
 }
 
-EllipticMotion& EllipticMotion::setRadius(float radius) {
+EllipticMotion& EllipticMotion::setRadius(double radius) {
     this->radius = radius;
     return *this;    
 }
 
-EllipticMotion& EllipticMotion::setAngle(float angle) {
+EllipticMotion& EllipticMotion::setAngle(double angle) {
     this->angle = angle;
     return *this;    
 }
 
-EllipticMotion& EllipticMotion::setAngularSpeed(float angularSpeed) {
+EllipticMotion& EllipticMotion::setAngularSpeed(double angularSpeed) {
     this->angularSpeed = angularSpeed;
     return *this;    
 }
 
-EllipticMotion& EllipticMotion::setEccentricity(float eccentricity) {
+EllipticMotion& EllipticMotion::setEccentricity(double eccentricity) {
     this->eccentricity = eccentricity;
     return *this;    
 }
 
-EllipticMotion& EllipticMotion::setEccentricityAngle(float eccentricityAngle) {
+EllipticMotion& EllipticMotion::setEccentricityAngle(double eccentricityAngle) {
     this->eccentricityAngle = eccentricityAngle;
     return *this;    
 }
@@ -67,13 +67,13 @@ EllipticMotion& EllipticMotion::setOverwrite(bool overwrite) {
 }
 
 void EllipticMotion::loop() {
-    float dT = timeElapsed / 1e6;
+    double dT = timeElapsed / 1e6;
     timeElapsed = 0;
 
-    float eccentricCompensation = (1 + cos(angle - eccentricityAngle) * eccentricity);
+    double eccentricCompensation = (1 + cos(angle - eccentricityAngle) * eccentricity);
     angle += angularSpeed * dT * eccentricCompensation;
 
-    float x = center + radius * cos(angle);
+    double x = center + radius * cos(angle);
     uint8_t brightness = MIN_BRIGHTNESS + (255 - MIN_BRIGHTNESS) * (sin(angle) + 1) / 2;
 
     pixel.setNormalized(x, CHSV(hue, saturation, brightness));
